compiler-pm-phase2/requirements.txt
# Python 3.10+
# ============================================

# Base & Utilities
python-dotenv==1.0.0

# Database (MySQL — Phase 1 backend)
mysql-connector-python>=8.2.0

# NLP & Parsing
sqlparse==0.4.4

# IA & LLM
openai>=1.3.9
ollama>=0.1.0

# Dashboard & Visualization
streamlit>=1.28.1
plotly>=5.17.0
pandas>=2.1.3
graphviz>=0.20.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Data Processing
numpy>=1.26.2

# PDF Professional Reports (§2.3)
reportlab>=4.0.0
matplotlib>=3.8.0
